+++ test/duplicate-steps/test.mjs
@@ -1,24 +1,49 @@
-import { normalize } from 'node:path';
-import { test, TestDir, execPlaywrightTestWithError, DEFAULT_CMD } from '../_helpers/index.mjs';
+import {
+  test,
+  TestDir,
+  normalize,
+  execPlaywrightTestWithError,
+  BDDGEN_CMD,
+} from '../_helpers/index.mjs';
 
 const testDir = new TestDir(import.meta);
+const featureFile = normalize('features/sample.feature');
 
-test(`${testDir.name} (main thread)`, () => {
-  execPlaywrightTestWithError(
-    testDir.name,
-    DUPLICATE_STEPS_ERROR,
-    `${DEFAULT_CMD} --project duplicates`,
-  );
+test(`${testDir.name} (main thread - regular steps)`, () => {
+  const error = [
+    `Multiple step definitions matched for text: "duplicate step" (${featureFile})`,
+    `  duplicate step - steps/steps.ts:6`,
+    `  duplicate step - steps/steps.ts:7`,
+    `  duplicate step - steps/steps.ts:8`,
+  ].join('\n');
+  execPlaywrightTestWithError(testDir.name, error, {
+    cmd: BDDGEN_CMD,
+    env: { PROJECTS: 'duplicate-regular-steps' },
+  });
 });
 
-test(`${testDir.name} (worker)`, () => {
-  execPlaywrightTestWithError(testDir.name, DUPLICATE_STEPS_ERROR);
+test(`${testDir.name} (main thread - decorator steps)`, () => {
+  const error = [
+    `Multiple step definitions matched for text: "duplicate decorator step" (${featureFile})`,
+    `  duplicate decorator step - steps/TodoPage.ts:7`,
+    `  duplicate decorator step - steps/TodoPage.ts:10`,
+    `  duplicate decorator step - steps/TodoPage.ts:13`,
+  ].join('\n');
+  execPlaywrightTestWithError(testDir.name, error, {
+    cmd: BDDGEN_CMD,
+    env: { PROJECTS: 'duplicate-decorator-steps' },
+  });
 });
 
-const DUPLICATE_STEPS_ERROR = [
-  `Multiple step definitions matched for text: "duplicate step" (${normalize(
-    'features/two.feature',
-  )})`,
-  '  duplicate step',
-  '  duplicate step',
-].join('\n');
+test(`${testDir.name} (worker - regular steps)`, () => {
+  const error = [
+    `Multiple step definitions matched for text: "duplicate step" (${featureFile})`,
+    `  duplicate step - steps/steps.ts:6`,
+    `  duplicate step - steps/steps.ts:7`,
+    `  duplicate step - steps/steps.ts:8`,
+  ].join('\n');
+  execPlaywrightTestWithError(testDir.name, error, {
+    cmd: BDDGEN_CMD,
+    env: { PROJECTS: 'no-duplicates,duplicate-regular-steps' },
+  });
+});